pyproject.toml
[project]
name = "stdl"
version = "0.5.2"
description = "Extended Python Standard Library"
authors = [{ name = "Žiga Ivanšek", email = "ziga.ivansek@gmail.com" }]
license = { file = "LICENSE.txt" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "PyYAML>=6.0",
    "tqdm",
    "python-dateutil>=2.8.2",
    "toml>=0.10.2 ",
]

classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

keywords = [
    "python utilites",
    "filesystem",
    "string manipulation",
    "logging configuration",
    "list utils",
    "standard library",
]


[project.optional-dependencies]
test = ["pytest"]
dev = [
    "black",
    "pytest",
    "ruff",
    "mkdocs",
    "mkdocs-material",
    "mkdocs-autorefs",
    "mkdocstrings[python]",
]

[project.urls]
Repository = "https://github.com/zigai/stdl"

[tool.black]
line-length = 100
target_version = ['py310']

[tool.ruff]
line-length = 100
extend-ignore = ["E731", "E741", "N802", "N803", "N806", "E501"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "E402"]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"
